+++ test/helpers/app-info.js
@@ -4,6 +4,7 @@ const appInfo = {
   appId: '4h2h4okwiyn8b6cle0oig00vitayum8ephrlsvg7xo8o19ne',
   appKey: '3xjj1qw91cr3ygjq9lt0g8c3qpet38rrxtwmmp0yffyoy2t4',
   masterKey: '3v7z633lzfec9qzx8sjql6zimvdpmtwypcchr2gelu5mrzb0',
+  hookKey: 'qlkcRqv9v5J5A11Byr3mzori',
   sessionTokenAdmin: 'w2jrtkbehp38otqmhbqu7ybs9'
 };